move all work with data to another module
