+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/dataprovider/dataprovider.iml" filepath="$PROJECT_DIR$/dataprovider/dataprovider.iml" />
       <module fileurl="file://$PROJECT_DIR$/yandex-mobilization.iml" filepath="$PROJECT_DIR$/yandex-mobilization.iml" />
     </modules>
   </component>
