+++ dataprovider/build.gradle
@@ -1,16 +1,18 @@
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 apply plugin: 'realm-android'
 apply plugin: 'me.tatarka.retrolambda'
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 24
+    buildToolsVersion "24.0.1"
 
     defaultConfig {
-        applicationId "com.volha.yandex.school.musicartists"
-        minSdkVersion 16
-        targetSdkVersion 23
+        minSdkVersion 11
+        targetSdkVersion 24
         versionCode 1
         versionName "1.0"
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
@@ -22,27 +24,15 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    dataBinding {
-        enabled = true
-    }
-    packagingOptions {
-        exclude 'META-INF/LICENSE'
-        exclude 'META-INF/NOTICE'
-    }
 }
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile 'com.android.support:design:23.1.1'
-
-    //Jackson
-    compile 'com.fasterxml.jackson.core:jackson-databind:2.7.0'
-    compile 'com.fasterxml.jackson.core:jackson-annotations:2.7.3'
-
-    //Image downloader
-    compile 'com.nostra13.universalimageloader:universal-image-loader:1.9.5'
-
+    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+        exclude group: 'com.android.support', module: 'support-annotations'
+    })
+    compile 'com.android.support:appcompat-v7:24.1.1'
+    testCompile 'junit:junit:4.12'
     //RxAndroid
     compile 'io.reactivex:rxandroid:1.0.1'
     compile 'io.reactivex:rxjava:1.0.14'
@@ -51,7 +41,4 @@ dependencies {
     compile 'com.squareup.retrofit:retrofit:2.0.0-beta1'
     compile 'com.squareup.retrofit:converter-jackson:2.0.0-beta1'
     compile 'com.squareup.retrofit:adapter-rxjava:2.0.0-beta1'
-
-    compile 'de.psdev.licensesdialog:licensesdialog:1.8.0'
-    compile project(path: ':dataprovider')
 }
